Refactor PR processing to improve clarity and consistency
